Extract NewTask into a hook
